Modified Vehicle

+++ logistics/src/main/java/org/bme/micro_futar/logistics/dtos/VehicleDTO.java
@@ -6,6 +6,6 @@ import lombok.Data;
 public class VehicleDTO {
     private Long id;
     private String registrationNumber;
-    private Long maximumPackableVolume;
+    private Double maximumPackableVolume;
 }
